Another Update!!
This is a new source folder to replace to one inside SimpleGameHome, a zip of the whole folder is to large to upload to github.

This Update has a new directory: src/TowerDefense
inside src/TowerDefense
The index.html document that has a temporary version of my game(a unity webgl player, there is a comment right above the div that contains it). There is an area above the game player for the title of the game and who the game is made by and another area below to display controls and any other additional information you want to display(optional).

The homepage also now links to this page.