Comment Section
This is a another change for the src folder in the SimpleGameHome folder.
I've reorganized the folder slightly and added a comment section to the page with the game player. The form works exactly like the one in Coffee Run I altered the names in it to work for an email, a comment, and a rating of 1 to 5 stars, unfortunately the posts do not save when the page refreshes.
I also restyled the color scheme and added more references in the scss/styles.scss file that can be altered easily if need be, but I think it looks nice.